moved uvm verification outside of src and refactored run.py into multifile modules
